Fix: Fix issue with generation of relationship constraints
When generating relationship constraints the constraint expressions on
the left hand side were not properly referencing the node objects, but
were instead somehow pulling in the value for the node. This led to an
inequality like "0 < constant * variable". This was addressed by
directly referencing the variable nodes when creating the constraint
expression.

This commit also implements a better print function for the optimizer
to show the scheduled providers and contacts after optimization.