Implement free solvers. Update README.
The repository was in a state it was setup to primarily work with
Gurobi, a commercial solver. Given that most people would not have
access to the costly license, this commit implements the hooks and
updates examples for using 3 different free solvers SCIP, Coin-OR Cbc,
and GLPK.

Additionally, it updates the documentation to add setup instructions
for all of these solves, removes the outdated Docker installation
instructions, and provides better usage instructions to reflect the
current state of the repository.